feat($digest): possibility to disable automatic digest cycle after a reception of data
This could be useful if you have yourself an obligation to execution a digest cycle in you app (if for example, you piped the ngStomp inside some Rx Observable). The builder now have a specific method `.withDigest(TRUE|FALSE)`. The builder has this value defined to true by default, for retro-compatibility and because it's the common use case.

Close #36