feat(connectionState): The state of the connection can be accessed by the promise on the ngStomp service
If you want to know if the system is connected or not, you can subscribe to the promise `ngStomp.connectionState.then()` and execute some code. `then` first function will be called if connection is up, second `then` function  or `catch` will be called if connection is down.

Don't forget, if you set a long timeOut in setting, the promise won't be fulfilled until we reach this timeOut. If the connection have already been open, the catch method won't be called

Linked to #43