feat(config): being able to configure the credentials / url after angular#config phase
You can now set login, password and url from the service with the simple setter.
The config has a new parameter, name `autoConnect` with default value to true. It allows to config if the system should connect when the service is created or simply initialise itself and wait an external call to ngStompService#connect

Close #44